Fix test cluster installation failure (#18)
One of worker nodes is used as bootstrap node. Docker is installed there
and then dhctl is used to install master. Then worker nodes are added to
the cluster as Static Instances. However, bootstrap node is failed to be
added to the cluster, because it already has containerd.

Add step to remove docker from bootstrap node after boostrap is done.